techpack: display: set doze brightness for aod correctly
Without this, the always on display uses much more power. It should probably be set from userspace
but setting it when the panel goes to LP2 mode means it always happens for AOD.
There is also a DOZE_BRIGHTNESS_LBM (low brightness mode) but that is very dim and doesn't use any less power.

Signed-off-by: palaych <palaychm@yandex.ru>